Moved recovery spec test under connectionRecoverySpec
