+++ src/IO.Ably.Tests.Shared/IO.Ably.Tests.Shared.projitems
@@ -100,7 +100,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\ConnectionSpecs\ConnectionParameterSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\ConnectionSpecs\ConnectionPingSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\ConnectionSpecs\ConnectionRecoverySpecs.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Realtime\ConnectionSpecs\ConnectionSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\ConnectionSpecs\EventEmitterSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\ConnectionSpecs\GeneralConnectionSpecs.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Realtime\ConnectionStateTests\ClosedStateSpecs.cs" />
