+++ src/IO.Ably.Tests.Shared/Realtime/ConnectionSpecs/ConnectionRecoverySpecs.cs
@@ -65,6 +65,27 @@ namespace IO.Ably.Tests.Realtime.ConnectionSpecs
             client.Connection.RecoveryKey.Should().Be(expectedRecoveryKey);
         }
 
+        [Fact]
+        [Trait("spec", "RTN16i")]
+        [Trait("spec", "RTN16f")]
+        [Trait("spec", "RTN16j")]
+        public async Task RecoveryKey_MsgSerialShouldNotBeSentToAblyButShouldBeSetOnConnection()
+        {
+            var recoveryKey =
+                "{\"connectionKey\":\"uniqueKey\",\"msgSerial\":45,\"channelSerials\":{\"channel1\":\"1\",\"channel2\":\"2\",\"channel3\":\"3\"}}";
+            FakeTransportFactory.InitialiseFakeTransport =
+                transport => transport.OnConnectChangeStateToConnected = false;
+            var client = GetClientWithFakeTransport(options => { options.Recover = recoveryKey; });
+
+            var transportParams = await client.ConnectionManager.CreateTransportParameters("https://realtime.ably.io");
+            var paramsDict = transportParams.GetParams();
+            paramsDict.ContainsKey("recover").Should().BeTrue();
+            paramsDict["recover"].Should().Be("uniqueKey");
+            paramsDict.ContainsKey("msg_serial").Should().BeFalse();
+            await Task.Delay(2000);
+            client.Connection.MessageSerial.Should().Be(45);
+        }
+
         public ConnectionRecoverySpecs(ITestOutputHelper output)
             : base(output)
         {
